requirements.txt
absl-py==2.3.1
astunparse==1.6.3
attrs==25.4.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dlib @ file:///D:/UCB/EMBEBIDOS_2/PROYECTO_FINAL/dlib-19.22.99-cp310-cp310-win_amd64.whl#sha256=f2181b6724669fb0147c9ebc326d72b94de41bced1812eff5df4b133b5d0b575
face-recognition==1.3.0
face_recognition_models==0.3.0
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.4.0
google-auth==2.43.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
jax==0.6.2
jaxlib==0.6.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
mediapipe==0.10.5
ml_dtypes==0.5.4
numpy==1.26.4
oauthlib==3.3.1
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pillow==12.0.0
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyparsing==3.2.5
pyserial==3.5
python-dateutil==2.9.0.post0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scipy==1.15.3
six==1.17.0
sounddevice==0.5.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.2.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==2.0.1
